requirements.txt
boto3==1.16.19
Django==3.1
djangorestframework==3.11.1
djangorestframework-api-key==2.*
rest-framework-generic-relations==2.0.0

django-crispy-forms
django-debug-toolbar~=3.2.3
git+https://github.com/makimo/django-eav2@master
django-filter~=21.1
django-multiselectfield
django-admin-rangefilter

psycopg2==2.8.4
python-decouple
numpy~=1.21.4

requests==2.22.0
Pillow
sentry-sdk==0.14.0

uvicorn[standard]
gunicorn==20.0.4

setuptools~=57.0.0
unicode~=2.8
lxml~=4.7.1
python-dateutil~=2.8.2

rollbar
django-rq==2.5.1
django-redis==5.2.0
django-countries==7.4.2
langdetect==1.0.9